Added particle collision with enemies that cause an explosion and despawns them Added a rigidbody gravity fall when the player dies Refactored some code

+++ Assets/Scripts/BehaviourHelpers.cs
@@ -9,12 +9,12 @@ namespace UnityEngine
     public static class BehaviourHelpers
     {
         // Instead of using Invoke and needing to use the name of your method, here's a helper
-        public static void DelayInvoke(MonoBehaviour behavior, Action action, int seconds)
+        public static void DelayInvoke(MonoBehaviour behavior, Action action, float seconds)
         {
             behavior.StartCoroutine(Wait(action, seconds));
         }
 
-        private static IEnumerator<YieldInstruction> Wait(Action action, int seconds)
+        private static IEnumerator<YieldInstruction> Wait(Action action, float seconds)
         {
             yield return new WaitForSeconds(seconds);
